Fix more warnings and adjust Cirrus CI rules
Change hts_close() to hts_close_abruptly() in bgzip, to fix an
empty body warning, allowing removal of -Wno-empty-body in CI
tests.

Fix signed compare in htslib/bgzf.h, which caused a warning in
the maintainer-check usepublic.cpp check.  It appears that -Wall
does not allow signed-compare in c++.

Make CFLAGS settings more consistent in .cirrus.yml, and ensure
$MAKE_OPTS is consistently applied in all make invocations.  The
latter is necessary because "make cc-version" and
"make maintainer-check" can trigger production of some .mk files,
and we want them to be made in the same way as if we ran the full
build.

Switch MacOS build to Sonoma as Ventura is no longer supported
by Cirrus.

+++ htslib/bgzf.h
@@ -182,7 +182,8 @@ static inline ssize_t bgzf_read_small(BGZF *fp, void *data, size_t length) {
  */
 static inline
 ssize_t bgzf_write_small(BGZF *fp, const void *data, size_t length) {
-    if (fp->is_compressed && BGZF_BLOCK_SIZE - fp->block_offset > length) {
+    if (fp->is_compressed
+        && (size_t) (BGZF_BLOCK_SIZE - fp->block_offset) > length) {
         // Short cut the common and easy mode
         memcpy((uint8_t *)fp->uncompressed_block + fp->block_offset,
                data, length);
